cleaned up prevalence rates for #99 reference to #73
