+++ app/config/database.php
@@ -55,9 +55,9 @@ return array(
 		'mysql' => array(
 			'driver'    => 'mysql',
 			'host'      => 'localhost',
-			'database'  => 'blis',
-			'username'  => 'root',
-			'password'  => 'Admin123',
+			'database'  => 'iblis',
+			'username'  => 'lara',
+			'password'  => 'password',
 			'charset'   => 'utf8',
 			'collation' => 'utf8_unicode_ci',
 			'prefix'    => '',
